Open external URI (#8)
* Fix tsc commands

They were complaining about the -p flag so I removed them since they do
not appear necessary.  I also removed --outDir since that is already
specified in the tsconfig.json.

* Uncapitalize workspace

It does not seem we capitalize it in our docs so this may be more
consistent.

* Handle external URIs

This makes it so we can have an "open in VS Code" button that
automatically launches our extension where it can handle downloading the
binary (todo) then opening the workspace (done).

* Use remote.SSH.configFile when set

Otherwise if it is set (either by the user or an extension like Gitpod)
our configuration will not be picked up since it is in the default
location.  VS Code will not read both the default location and this
location.