refactor: update logic for downloading binary
This makes significant changes to the `fetchBinary` logic.

First, it refactors a couple pieces of logic into methods on the
`Storage` class to make the code more readable.

Then it modifies the flow to first check if the binary is outdated. If
it is, then it downloads the latest version.