Check that returned user has roles
You can get undefined roles when the request succeeds but does not
return an actual user (like if you are trying to break the extension by
testing it with a non-Coder domain and you get a 200 and HTML back).
That might be a bug with the SDK actually but for now fix it here.