Mark untrusted workspaces as supported
It seems that if you do not use any files in the workspace then you can
set this as true.

This fixes an issue where you get "no view is registered with id"
because the extension seems to be in a half-activated state where the
activate function runs but the view does not exist.  Probably a bug in
VS Code but we might as well support untrusted workspaces anyway.