Ensure we have a URL
It looks like when we switched to using the VS Code desktop module the
URL was no longer included, which caused the extension to query
localhost if you had never logged in before, which errors and causes the
configuration to never happen.

I will fix the module as well, but I figure we can check the URL in the
plugin just in case.