Fix error when agent has no metadata
The problem is that we dispose the watcher, which removes it from the
array, but then immediately try to get it off the array.  It then throws
into the catch where it tries to get it off the array again to
dispose (again).

Use a local object instead of getting it off the array each time.