Upgrade to latest Coder
We need this to create standalone clients so we can support multiple
deployments.  Currently the plugin always uses the url and token for the
current deployment, but you might be trying to launch a previous
workspace that belongs to a different deployment (from the recents
projects menu for example).

Also, lots of the code relies on the url and token being stored and
accessed globally (storage.getURL(), storage.getSessionToken()) so avoid
those as much as possible (only use when creating a client, basically).