Scope SSH configs by label
This lets us create multiple ssh config blocks that each point to
different Coder binaries and config directories for those binaries,
instead of always overwriting a single configuration block, binary, and
config directory with the most recently configured deployment is like we
currently do.

This scoping is accomplished by changing the ssh config block's comment
markers to include the label and by adding the label to the host name:
coder-vscode.label--*.  The Coder binary splits by -- and ignores the
first part, so we are free to use it however we need.

This scoping is used support config blocks per deployment, where the
label is derived from the deployment URL.  That means we can have one
config block for dev.coder.com, another for test.coder.com, and so on,
each with their own binaries and tokens, coexisting in harmony.