Read url and token from cli config
Instead of pulling the url and token from the currently logged-in
deployment, we use the url and token as they exist on the disk organized
by the label (~/.config/.../coder.coder-remote/dev.coder.com for
example).

The label we extract from the host (coder-vscode.dev.coder.com-- for
example).  Because the new format is a little different, the parsing has
also changed.  It has been moved to utils for easier testing.

This way, we pull the right url and token for the host even if the user
is logged into a different deployment currently.

If there is no label because it is the older format, we read the old
deployment-unaware path instead (~/.config/.../coder.coder-remote).

Since this now uses the same url and token the cli itself will use, this
has the extra advantage of preventing a potential desync between what
exists on disk and what exists in VS Code's memory.