Remove coder.openFromSidebar from command list
When searching for "open workspace" in the palette, this can come up
instead of the correct one (`coder.open`, also titled "open workspace").

coder.openFromSidebar should only be used to open workspaces from the
sidebar, it never makes sense for a user to manually run it from the
command palette because even if you have a tree item selected, it will
not be passed when running from the command palette; the tree item only
exists if you click the inline button.

If there was a way to get the selected tree item, it would be ideal to
enable the command while there was one (this would also allow adding a
key binding), but since there is not, use `false` to always hide it.