Use web socket for workspace notifications
Previously, we were polling all the user's workspaces to get information
about autostop and deletion so we can notify the user when those things
are about to happen.

This has a few problems:

1. We really only want to notify for the workspace to which we are
   actively connected, not all workspaces.
2. It did not account for owners being connected to someone else's
   workspace.  We would have to query all workspaces, which is an even
   more expensive query.
3. If the sidebar is open, we are running two of the same query every
   five seconds, which is wasteful.

We already had a web socket that was used to notify about the workspace
being outdated, so I broke that into a new class and combined all the
notification (outdated, autostop, deletion), status bar updates (just
shows if the workspace is outdated), and context updates into one place
using that web socket.