Tweak update text
It says the window will reload, but the command has no way of knowing
that is the case.  Maybe the command could automatically reload, but for
now, a second notification will inform the user that the window needs to
be reloaded, and they can choose whether they want to do so.