Remove agent singleton
This causes TLS certificates to be reread on every API request, so if
they are refreshed on disk, the extension will pick up the new ones.
It also avoids the need to recreating the REST client (which was an
ultimately unsuccessful effort to make the extension pick up new
certificates).