Fix connecting to off workspaces
We made a change to always include the agent in the host name so the
session is stable, however for workspaces that are off there is no agent
included in the response, and we were silently returning in these cases.

So for these cases, fetch the first agent through a different API that
still returns agents when the workspace is off.

Workspace picker is broken out into a separate function, logic is
unchanged.

Instead of a possibly-undefined agent on a tree item, I switch on the
tree item type.  I feel the intent this way is easier to follow.

To extract agents from both APIs, make extractAgents accept an array of
resources instead of a workspace.